+++ resources/lang/de/validation.php
@@ -266,4 +266,7 @@ return [
         'restrict'                          => 'Verwendung einschränken',
         'room_token_expiration'             => 'Ablaufzeit für personalisierte Raumlinks'
     ],
+
+    'validname'       => 'Der Name enthält die folgenden nicht erlaubten Zeichen: :chars',
+    'validname_error' => 'Der Name enthält nicht erlaubten Zeichen'
 ];